Implement Cart and CartService with add/remove functionality; enhance UserEntity relationships and refactor total price calculation

+++ src/main/java/com/soulittude/e_commerce/entity/Cart.java
@@ -30,8 +30,9 @@ public class Cart {
     @PreUpdate
     public void calculateTotal() {
         this.totalPrice = items.stream()
-                .map(item -> item.getProduct().getPrice()
-                        .multiply(BigDecimal.valueOf(item.getQuantity())))
-                .reduce(BigDecimal.ZERO, BigDecimal::add);
+            .map(item -> 
+                BigDecimal.valueOf(item.getProduct().getPrice()) // Convert Double to BigDecimal
+                .multiply(BigDecimal.valueOf(item.getQuantity()))
+            .reduce(BigDecimal.ZERO, BigDecimal::add));
     }
 }
